add AppStore collapse/expand filters working
