+++ packages/frontend/src/components/facets/utils.ts
@@ -182,12 +182,13 @@ export const useUpdateFilters = (index: string) => {
     );
   };
 };
-export const useGetFacetFilters = (
-  index: string,
-  field: string,
-): Operation | undefined => {
-  return useCoreSelector((state: CoreState) =>
-    selectIndexedFilterByName(state, index, field),
+export const useGetFacetFilters = (index: string, field: string): Operation => {
+  return useCoreSelector(
+    (state: CoreState) =>
+      selectIndexedFilterByName(state, index, field) ?? {
+        operator: 'and',
+        operands: [],
+      },
   );
 };
